requirements.txt
# python -m venv myenv

# myenv\Scripts\Activate

streamlit
google-generativeai
python-dotenv
